+++ src/snowflake/snowpark/modin/plugin/compiler/snowflake_query_compiler.py
@@ -16580,7 +16580,12 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
             )
         }
 
-        rows = join_result.result_frame.ordered_dataframe.agg(agg_exprs).collect()
+        try:
+            rows = join_result.result_frame.ordered_dataframe.agg(agg_exprs).collect()
+        except Exception:
+            # return not equal for any exception from SQL, e.g., type conversion error while comparing with two
+            # different types.
+            return False
         # In case of empty table/dataframe booland_agg returns None. Add special case
         # handling for that.
         return all(x is None for x in rows[0]) or all(rows[0])